refactor: change the way to load shell
