--- include/zen-shell/zen-shell.h
@@ -1,8 +1,0 @@
-#ifndef ZEN_SHELL_H
-#define ZEN_SHELL_H
-
-#include <libzen-compositor/libzen-compositor.h>
-
-int zen_shell_init(struct zen_compositor *compositor);
-
-#endif  //  ZEN_SHELL_H
