+++ tests/meson.build
@@ -1,7 +1,11 @@
 tests = {
   'noop-test': {},
   'load-backend': {
-    'deps': [ dep_libzen_compositor_private, dep_headless_backend_public ],
+    'deps': [
+      dep_libzen_compositor_private,
+      dep_headless_backend_public,
+      dep_zen_shell_public
+    ],
   },
 }
