+++ zen-backend/glfw/output.c
@@ -1,3 +1,5 @@
+#include "output.h"
+
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
 #include <libzen-compositor/libzen-compositor.h>
