+++ zen-shell/shell.c
@@ -1,15 +1,12 @@
 #include <libzen-compositor/libzen-compositor.h>
 #include <wayland-server.h>
-#include <zen-shell/zen-shell.h>
 #include <zigen-shell-server-protocol.h>
 
 struct zen_shell {
   struct zen_compositor* compositor;
 
-  struct wl_listener compositor_destroy_listener;
+  struct wl_global* global;
 };
-
-static void zen_shell_destroy(struct zen_shell* shell);
 
 static void
 zen_shell_protocol_destroy(
@@ -37,17 +34,6 @@ static const struct zgn_shell_interface shell_interface = {
 };
 
 static void
-zen_shell_compositor_destroy_handler(struct wl_listener* listener, void* data)
-{
-  UNUSED(data);
-  struct zen_shell* shell;
-
-  shell = wl_container_of(listener, shell, compositor_destroy_listener);
-
-  zen_shell_destroy(shell);
-}
-
-static void
 zen_shell_bind(
     struct wl_client* client, void* data, uint32_t version, uint32_t id)
 {
@@ -64,7 +50,7 @@ zen_shell_bind(
   wl_resource_set_implementation(resource, &shell_interface, shell, NULL);
 }
 
-static struct zen_shell*
+WL_EXPORT struct zen_shell*
 zen_shell_create(struct zen_compositor* compositor)
 {
   struct zen_shell* shell;
@@ -84,10 +70,7 @@ zen_shell_create(struct zen_compositor* compositor)
   }
 
   shell->compositor = compositor;
-  shell->compositor_destroy_listener.notify =
-      zen_shell_compositor_destroy_handler;
-  wl_signal_add(
-      &compositor->destroy_signal, &shell->compositor_destroy_listener);
+  shell->global = global;
 
   return shell;
 
@@ -98,23 +81,9 @@ err:
   return NULL;
 }
 
-static void
+WL_EXPORT void
 zen_shell_destroy(struct zen_shell* shell)
 {
-  wl_list_remove(&shell->compositor_destroy_listener.link);
+  wl_global_destroy(shell->global);
   free(shell);
-}
-
-WL_EXPORT int
-zen_shell_init(struct zen_compositor* compositor)
-{
-  struct zen_shell* shell;
-
-  shell = zen_shell_create(compositor);
-  if (shell == NULL) {
-    zen_log("shell: failed to create zen_shell\n");
-    return -1;
-  }
-
-  return 0;
 }
