+++ zen/main.c
@@ -1,6 +1,5 @@
 #include <libzen-compositor/libzen-compositor.h>
 #include <signal.h>
-#include <zen-shell/zen-shell.h>
 
 static int
 on_term_signal(int signal_number, void *data)
@@ -33,15 +32,16 @@ main()
     goto out_compositor;
   }
 
+  ret = zen_compositor_load_shell(compositor);
+  if (ret != 0) {
+    zen_log("main: failed to load shell\n");
+    goto out_shell;
+  }
+
   ret = zen_compositor_load_backend(compositor);
   if (ret != 0) {
     zen_log("main: failed to load backend\n");
     goto out_backend;
-  }
-
-  if (zen_shell_init(compositor) != 0) {
-    zen_log("main: failed to load shell\n");
-    goto out_shell;
   }
 
   ret = wl_display_add_socket(display, socket);
@@ -66,8 +66,8 @@ out_signal:
     if (signals[i]) wl_event_source_remove(signals[i]);
 
 out_socket:
-out_shell:
 out_backend:
+out_shell:
   zen_compositor_destroy(compositor);
 
 out_compositor:
